Prepare a 4.2-M1 release of tracing
Prepare a 4.2-M1 release of tracing, by firstly updating the Axon
Framework version to 4.2. Secondly, the gateway implementations should
be converted to a wrapper instead of an extension of the existing
gateways. This will make using the Tracing solution more lenient and
more importantly ensure the new way of dealing with exceptions in a
distributed mechanism is correctly supported.