Move Flux tracing to utils class
Any reactor calls in the body of the TracingQueryGateway class will lead to a ClassNotFoundException upon boot of the application. The JVM will read the class, encounter a project reactor class and break.

This PR will remove this dependency in the body of the class and move it to a class that is not inspected using reflection: The SpanUtils